+++ lib/ui/emulator/debugger/breakpoint_dialog.dart
@@ -4,9 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:nesd/extension/hex_extension.dart';
+import 'package:nesd/nes/debugger/breakpoint.dart';
 import 'package:nesd/nes/debugger/debugger.dart';
 import 'package:nesd/nes/debugger/debugger_state.dart';
-import 'package:nesd/nes/nes.dart';
 import 'package:nesd/ui/emulator/debugger/debugger_widget.dart';
 
 class BreakpointDialog extends ConsumerWidget {